// src/services/proxy/generated/models/LineaPedido.ts
/* tslint:disable */
/* eslint-disable */
/**
 * The Golden Book API
 * API para negocio de venta de libros 
 *
 * The version of the OpenAPI document: 1.0
 * Contact: soporte@thegoldenbook.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface LineaPedido
 */
export interface LineaPedido {
    /**
     * 
     * @type {number}
     * @memberof LineaPedido
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof LineaPedido
     */
    precio?: number;
    /**
     * 
     * @type {number}
     * @memberof LineaPedido
     */
    libroId?: number;
    /**
     * 
     * @type {string}
     * @memberof LineaPedido
     */
    nombreLibro?: string;
    /**
     * 
     * @type {number}
     * @memberof LineaPedido
     */
    pedidoId?: number;
    /**
     * 
     * @type {number}
     * @memberof LineaPedido
     */
    unidades?: number;
}

/**
 * Check if a given object implements the LineaPedido interface.
 */
export function instanceOfLineaPedido(value: object): value is LineaPedido {
    return true;
}

export function LineaPedidoFromJSON(json: any): LineaPedido {
    return LineaPedidoFromJSONTyped(json, false);
}

export function LineaPedidoFromJSONTyped(json: any, ignoreDiscriminator: boolean): LineaPedido {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'precio': json['precio'] == null ? undefined : json['precio'],
        'libroId': json['libroId'] == null ? undefined : json['libroId'],
        'nombreLibro': json['nombreLibro'] == null ? undefined : json['nombreLibro'],
        'pedidoId': json['pedidoId'] == null ? undefined : json['pedidoId'],
        'unidades': json['unidades'] == null ? undefined : json['unidades'],
    };
}

export function LineaPedidoToJSON(json: any): LineaPedido {
    return LineaPedidoToJSONTyped(json, false);
}

export function LineaPedidoToJSONTyped(value?: LineaPedido | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'precio': value['precio'],
        'libroId': value['libroId'],
        'nombreLibro': value['nombreLibro'],
        'pedidoId': value['pedidoId'],
        'unidades': value['unidades'],
    };
}